requirements/base.txt
Django==3.2  # Use LTS version
psycopg2==2.9.2 # Python-PostgreSQL Database Adapter
celery==5.2.1 # The latest version for Python3.8
requests==2.28.1 # An HTTP library for Python
redis==3.5.3  # Celery broker, make sure it matches celery
SQLAlchemy==1.4.28  # The Database Toolkit for Python
kobo==0.20.1
filetype==1.0.8
extractcode==30.0.0
scancode-toolkit==30.1.0
extractcode-7z==16.5.210531
extractcode-libarchive==3.5.1.210531
# FIXME: latest version of workflow only supports python3.6
workflow==2.1.6
koji==1.27.0
djangorestframework==3.13.1
django-filter==21.1
Markdown==3.3.6
swh.model==3.2.0
jsonschema==4.3.3
gunicorn==20.1.0
krbcontext==0.10
python-ldap==3.4.0
ijson==3.1.4
httpx==0.23.0
uvloop==0.16.0
django-mptt==0.13.4
